fix(daemon): contain deferred worker retries
