+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
-  "maximum_collected_cases": 16040,
-  "maximum_test_source_lines": 346861
+  "maximum_collected_cases": 16041,
+  "maximum_test_source_lines": 346884
 }
